refactor: rafactor

+++ backend/internal/adapter/repository/bus_repo.go
@@ -3,6 +3,7 @@ package repository
 import (
 	"database/sql"
 	"errors"
+	"strings"
 	"time"
 	"unitrip/internal/entity"
 	"unitrip/internal/usecase"
@@ -19,6 +20,10 @@ func NewBusRepo(db *sql.DB) usecase.BusRepo {
 }
 
 func (b *bus) GetAll() ([]*entity.Bus, error) {
+	if b.db == nil {
+		return nil, errors.New("database connection is nil")
+	}
+
 	query := `
 		SELECT
 			b.id,
@@ -27,11 +32,14 @@ func (b *bus) GetAll() ([]*entity.Bus, error) {
 			b.travel_date,
 			b.travel_time,
 			b.price,
+			b.rating_avg,
+			b.rating_count,
 			c1.id, c1.name,
 			c2.id, c2.name
 		FROM buses b
 		JOIN cities c1 ON c1.id = b.source_city_id
 		JOIN cities c2 ON c2.id = b.dest_city_id
+		ORDER BY b.id DESC
 	`
 
 	rows, err := b.db.Query(query)
@@ -46,6 +54,7 @@ func (b *bus) GetAll() ([]*entity.Bus, error) {
 		var bus entity.Bus
 		var source entity.City
 		var dest entity.City
+		var rating entity.Rating
 
 		var dateStr, timeStr string
 
@@ -56,6 +65,8 @@ func (b *bus) GetAll() ([]*entity.Bus, error) {
 			&dateStr,
 			&timeStr,
 			&bus.Price,
+			&rating.Average,
+			&rating.Count,
 			&source.ID, &source.Name,
 			&dest.ID, &dest.Name,
 		)
@@ -63,16 +74,20 @@ func (b *bus) GetAll() ([]*entity.Bus, error) {
 			return nil, err
 		}
 
+		dateStr = strings.Split(dateStr, "T")[0]
+		timeStr = strings.TrimSuffix(strings.Split(timeStr, "T")[1], "Z")
+
 		dt, _ := time.Parse(
 			"2006-01-02 15:04:05",
 			dateStr+" "+timeStr,
 		)
 		bus.Schedule = dt
 
-		bus.Route = entity.Route{
-			Source: source,
-			Dest:   dest,
+		bus.Route = &entity.Route{
+			Source: &source,
+			Dest:   &dest,
 		}
+		bus.Rating = &rating
 
 		buses = append(buses, &bus)
 	}
@@ -81,12 +96,18 @@ func (b *bus) GetAll() ([]*entity.Bus, error) {
 }
 
 func (b *bus) GetByID(id int64) (*entity.Bus, error) {
+	if b.db == nil {
+		return nil, errors.New("database connection is nil")
+	}
+
 	query := `
 		SELECT
 			b.id,
 			b.travel_date,
 			b.travel_time,
 			b.price,
+			b.rating_avg,
+			b.rating_count,
 			c1.id, c1.name,
 			c2.id, c2.name
 		FROM buses b
@@ -98,6 +119,7 @@ func (b *bus) GetByID(id int64) (*entity.Bus, error) {
 	var bus entity.Bus
 	var source entity.City
 	var dest entity.City
+	var rating entity.Rating
 
 	var dateStr, timeStr string
 
@@ -106,9 +128,14 @@ func (b *bus) GetByID(id int64) (*entity.Bus, error) {
 		&dateStr,
 		&timeStr,
 		&bus.Price,
+		&rating.Average,
+		&rating.Count,
 		&source.ID, &source.Name,
 		&dest.ID, &dest.Name,
 	)
+
+	dateStr = strings.Split(dateStr, "T")[0]
+	timeStr = strings.TrimSuffix(strings.Split(timeStr, "T")[1], "Z")
 
 	if err != nil {
 		switch {
@@ -125,15 +152,20 @@ func (b *bus) GetByID(id int64) (*entity.Bus, error) {
 	)
 	bus.Schedule = dt
 
-	bus.Route = entity.Route{
-		Source: source,
-		Dest:   dest,
+	bus.Route = &entity.Route{
+		Source: &source,
+		Dest:   &dest,
 	}
+	bus.Rating = &rating
 
 	return &bus, nil
 }
 
 func (b *bus) Create(bus *entity.Bus) (*entity.Bus, error) {
+	if b.db == nil {
+		return nil, errors.New("database connection is nil")
+	}
+	
 	query := `
 		INSERT INTO buses (
 			source_city_id,
@@ -165,7 +197,11 @@ func (b *bus) Create(bus *entity.Bus) (*entity.Bus, error) {
 	return bus, nil
 }
 
-func (r *bus) Update(bus *entity.Bus) error {
+func (b *bus) Update(bus *entity.Bus) error {
+	if b.db == nil {
+		return errors.New("database connection is nil")
+	}
+
 	query := `
 		UPDATE buses
 		SET
@@ -180,7 +216,7 @@ func (r *bus) Update(bus *entity.Bus) error {
 	date := bus.Schedule.Format("2006-01-02")
 	time := bus.Schedule.Format("15:04:05")
 
-	_, err := r.db.Exec(
+	_, err := b.db.Exec(
 		query,
 		bus.Route.Source.ID,
 		bus.Route.Dest.ID,
@@ -194,12 +230,20 @@ func (r *bus) Update(bus *entity.Bus) error {
 }
 
 func (b *bus) Delete(id int64) error {
+	if b.db == nil {
+		return errors.New("database connection is nil")
+	}
+
 	query := `DELETE FROM buses WHERE id = $1`
 	_, err := b.db.Exec(query, id)
 	return err
 }
 
 func (b *bus) GenerateSeats(id int64, n int) error {
+	if b.db == nil {
+		return errors.New("database connection is nil")
+	}
+
 	query := `
 		INSERT INTO bus_seats (bus_id, seat_no)
 		SELECT $1, generate_series(1, $2)
@@ -209,8 +253,12 @@ func (b *bus) GenerateSeats(id int64, n int) error {
 }
 
 func (b *bus) GetSeatsByID(id int64) ([]*entity.Seat, error) {
+	if b.db == nil {
+		return nil, errors.New("database connection is nil")
+	}
+
 	query := `
-		SELECT id, bus_id, seat_no, status
+		SELECT seat_no, status
 		FROM bus_seats
 		WHERE bus_id = $1
 		ORDER BY seat_no
